// types/profilesdefs.ts
export interface LockerSlot {
  items: string[];
  activeVariants?: {
    variants: any[];
  }[];
}

export type LockerSlotKey =
  | "Backpack"
  | "Character"
  | "Dance"
  | "Glider"
  | "ItemWrap"
  | "LoadingScreen"
  | "MusicPack"
  | "Pickaxe"
  | "SkyDiveContrail";

export interface IProfile {
  _id: string;
  createdAt: string;
  updatedAt: string;
  rvn: number;
  wipeNumber: number;
  accountId: string;
  profileId: string;
  version: string;
  stats: {
    templateId?: string;
    attributes: Partial<StatsAttributes>;
  };
  items: {
    [key: string]: {
      templateId: string;
      attributes: Partial<ItemValue>;
      quantity: number;
    };
  };
  commandRevision: number;
}
// ...
export interface Variants {
  channel: string;
  active: string;
  owned: string[];
}

export interface Lootlist {
  itemType: string;
  itemGuid: string;
  itemProfile: string;
  quantity: number;
}

interface Gifts {
  templateId: string;
  attributes: {
    lootList: Lootlist[];
  };
  quantity: number;
}

interface SavedRecords {
  recordIndex: number;
  archiveNumber: number;
  recordFilename: string;
}

interface PlacedBuildings {
  buildingTag: string;
  placedTag: string;
}

interface ProgressionInfo {
  progressionLayoutGuid: string;
  highestDefeatedTier: number;
}

interface GiftParameters {
  userMessage: string;
}

interface ObjectiveState {
  Name?: string;
  Value?: number;
  BackendName?: string;
  Count?: number;
  Stage?: number;
}

export interface ItemValue {
  platform: string;
  gender?: string;
  personality?: string;
  fromAccountId?: string;
  time?: string;
  squad_slot_idx?: number;
  userMessage?: string;
  templateIdHashed?: string;
  quest_state?: string;
  portrait?: string;
  building_slot_used?: number;
  set_bonus?: string;
  [key: string]: any;
  sent_new_notification: boolean;
  ObjectiveState: ObjectiveState[];
  lootList?: any[];
  alterationDefinitions?: any[];
  baseClipSize?: number;
  item_seen: boolean;
  refundsUsed?: number;
  refundCredits?: number;
  params: GiftParameters;
  sectionStates?: any[];
  state?: string;
  cloud_save_info?: {
    saveCount: number;
    savedRecords: SavedRecords[];
  };
  outpost_core_info?: {
    placedBuildings: PlacedBuildings[];
    accountsWithEditPermission: any[];
    highestEnduranceWaveReached: string;
  };
  tier_progression?: {
    progressionInfo: ProgressionInfo[];
  };
  level: number;
  max_level_bonus: number;
  rnd_sel_cnt: number;
  legacy_alterations?: any[];
  refund_legacy_item?: boolean;
  alterations?: any[];
  refundable?: boolean;
  alteration_base_rarities?: any[];
  homebaseBannerColorId?: string;
  loadedAmmo?: number;
  durability?: number;
  inventory_overflow_date?: boolean;
  itemSource?: string;
  quest_complete_playsolo?: number;
  quest_unlock_researchgadgets?: number;
  current_mtx_currency?: number;
  squad_slots_unlocked?: number;
  quest_unlock_buildweapons?: number;
  current_fort_profile_banner?: string;
  quest_unlock_defenders?: number;
  commander_level?: number;
  time_played?: number;
  squad_slot_count?: number;
  quest_unlock_eventquest?: number;
  quest_unlock_missiondefender?: number;
  homebaseBannerIconId?: string;
  squad_slot_count_1?: number;
  accountLevel: number;
  squad_slot_count_2?: number;
  daily_rewards_claimed?: number;
  seasonal_gold?: number;
  quest_unlock_craftweapons?: number;
  quest_unlock_personalassistant?: number;
  quest_unlock_upgradeweapon?: number;
  book_level: number;
  book_xp: number;
  book_sections_completed?: number;
  book_section_idx?: number;

  xp_overflow: number;
  quest_unlock_missionevent: number;
  purchased_slots: number;
  squad_slot_count_3: number;
  squad_slot_count_4: number;
  homebaseBannerColorIdBattleRoyal: string;
  currentSeason: number;
  homebaseName: string;
  variants: Variants[];
  xp: number;
  purchases?: any[];
  undo_cooldowns?: any[];
  mtx_affiliate_set_time?: string;
  inventory_limit_bonus?: number;
  mtx_affiliate?: string;
  forced_intro_played?: string;
  weekly_purchases?: any;
  daily_purchases?: any;
  ban_history?: {
    banCount: number;
    banTier: any;
  };
  in_app_purchases?: any;
  permissions?: any[];
  undo_timeout?: string;
  monthly_purchases?: any;
  allowed_to_send_gifts?: boolean;
  mfa_enabled?: boolean;
  unlock_epoch?: string;
  granted_bundles?: string[];
  has_unlock_by_completion?: boolean;
  num_quests_completed?: number;
  max_allowed_bundle_level?: number;
  num_granted_bundle_quests?: number;
  challenge_bundle_schedule_id?: string;
  num_progress_quests_completed?: number;
  grantedquestinstanceids?: string[];
  allowed_to_receive_gifts?: boolean;
  gift_history?: any;
  gifts?: Gifts[];
  banner_color?: string;
  homebase_name?: string;
  banner_icon?: string;
  ban_status?: {
    bRequiresUserAck: boolean;
    bBanHasStarted: boolean;
    banReasons: any[];
    banStartTimeUtc: string | null;
    banDurationDays: number | null;
    additionalInfo: string;
    exploitProgramName: string;
    competitiveBanReason: string;
  };
  survey_data?: any;
  personal_offers?: any;
  intro_game_played?: boolean;
  import_friends_claimed?: any;
  mtx_purchase_history?: {
    refundsUsed: number;
    refundCredits: number;
    purchases: any[];
  };
  locker_slots_data: {
    slots: {
      [slotName: string]: {
        items: string[];
        activeVariants?: any[];
      };
    };
  };
  use_count: number;
  banner_icon_template: string;
  banner_color_template: string;
  locker_name: string;
  favorite: boolean;
}